+++ shenyu-admin/src/test/java/org/apache/shenyu/admin/AbstractReflectGetterSetterTest.java
@@ -27,7 +27,6 @@ import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.math.BigDecimal;
 import java.sql.Timestamp;
-import java.util.Arrays;
 import java.util.Collections;
 import java.util.Date;
 import java.util.HashMap;
@@ -88,7 +87,7 @@ public abstract class AbstractReflectGetterSetterTest {
     protected abstract Class<?> getTargetClass();
 
     protected Set<String> getExcludeFields() {
-        return new HashSet<>(Arrays.asList("serialVersionUID"));
+        return new HashSet<>(Collections.singletonList("serialVersionUID"));
     }
 
     /**
@@ -97,6 +96,7 @@ public abstract class AbstractReflectGetterSetterTest {
      * @throws Exception maybe throw reflect Exception.
      */
     @Test
+    @SuppressWarnings("all")
     public void testGetAndSet() throws Exception {
         Class<?> clazz = getTargetClass();
         Object target = clazz.getDeclaredConstructor().newInstance();